16-bit Mesh Quantization

+++ Engine/shaders/glsl/3D.vert
@@ -9,7 +9,7 @@
 #define MAX_MATRICES 20
 #endif
 
-layout(location = 0) in vec3 i_pos;
+layout(location = 0) in uvec3 i_pos;
 layout(location = 1) in vec3 i_normal;
 layout(location = 2) in vec2 i_uv;
 layout(location = 3) in int tex_sub_index;
